+++ src/buildTree.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode'
 import * as path from 'path'
-import { currentBmx, isPartOfWorkspace } from './common'
+import { currentBmx } from './common'
 import { currentDefinition } from './taskProvider'
 
 export class BmxBuildTreeProvider implements vscode.TreeDataProvider<vscode.TreeItem> {
@@ -80,8 +80,8 @@ export class BmxBuildTreeProvider implements vscode.TreeDataProvider<vscode.Tree
 				this.isForWorkspace = workspaceFolder ? true : false
 			} else {
 				
-				if (vscode.workspace && vscode.workspace.getWorkspaceFolder.length > 0){
-					rootName = 'Workspace: ' + vscode.workspace.name?.toUpperCase()
+				if (vscode.workspace && vscode.workspace.getWorkspaceFolder.length > 0 && vscode.workspace.name){
+					rootName = 'Workspace: ' + vscode.workspace.name.toUpperCase()
 					this.isForWorkspace = true
 				} else {
 					this.isForWorkspace = false